+++ Helpers/Bank.lua
@@ -5,13 +5,18 @@ RS.didBankStuff = false
 RS.justSplit = false
 RS.splitLoc = {}
 
+local BACKPACK_CONTAINER = 0
+local BANK_CONTAINER = -1
+local PLAYER_BAGS = {}
+local PLAYER_BAGS_REVERSED = {}
+
 local BANK_BAGS -- set up in RS.SetupBankConstants
 local BANK_BAGS_REVERSED -- set up in RS.SetupBankConstants
 
-if REAGENTBANK_CONTAINER then
-  tinsert(BANK_BAGS, REAGENTBANK_CONTAINER)
-  tinsert(BANK_BAGS_REVERSED, REAGENTBANK_CONTAINER)
-end
+--if REAGENTBANK_CONTAINER then
+--  tinsert(BANK_BAGS, REAGENTBANK_CONTAINER)
+--  tinsert(BANK_BAGS_REVERSED, REAGENTBANK_CONTAINER)
+--end
 
 --local GetContainerItemInfo = _G.GetContainerItemInfo
 
@@ -25,6 +30,9 @@ function RS.SetupBankConstants()
     BANK_BAGS = { BANK_CONTAINER, 5, 6, 7, 8, 9, 10 }
     BANK_BAGS_REVERSED = { 10, 9, 8, 7, 6, 5, BANK_CONTAINER }
   end
+
+  PLAYER_BAGS = { 0, 1, 2, 3, 4 }
+  PLAYER_BAGS_REVERSED = { 4, 3, 2, 1, 0 }
 end
 
 local function rsCount(T)
@@ -37,7 +45,7 @@ local function rsCount(T)
 end
 
 local function rsIsSomethingLocked()
-  for bag = BACKPACK_CONTAINER, NUM_BAG_SLOTS do
+  for _, bag in ipairs(PLAYER_BAGS) do
     for slot = 1, GetContainerNumSlots(bag) do
       local _, _, locked = GetContainerItemInfo(bag, slot)
       if locked then
@@ -123,221 +131,115 @@ local function rsGetItemsInBank()
   return result
 end
 
----@param item RsItem
----@param amountOnMouse number
-local function rsPutSplitItemIntoBags(item, amountOnMouse)
+---@param dropItem RsItem
+---@param bag number BagID https://wowwiki-archive.fandom.com/wiki/BagId
+---@return number, number Returns bag, slot where drop happened
+local function rsDropCursorItemIntoBag(dropItem, bag)
+  -- Search through the bag for an empty slot
+  for slot = 1, GetContainerNumSlots(bag) do
+    local _, itemCount, locked, _, _, _, _, _, _, itemID = GetContainerItemInfo(bag, slot)
+
+    if not locked and not itemCount then
+      PickupContainerItem(bag, slot)
+      return bag, slot
+    end
+  end -- for all bag slots
+  return nil
+end
+
+---Takes cursor item. Drops it into same bags where it was taken from then right-clicks it to move
+---into the opposite bags group. I.e. bank to bags, or bags to bank.
+---@param dropItem RsItem
+---@param srcBags table<number, number> Bag list where the item comes from - used for splitting
+---@return number, number Bag and slot where the drop happened
+local function rsSplitSwapCursorItem(dropItem, srcBags)
   if not CursorHasItem() then
     return
   end
   C_NewItems.ClearAll()
 
-  -- for all bags: backpack, 1,2,3,4
-  -- Try find an incomplete stack that will fit the items we're moving
-  for bag = BACKPACK_CONTAINER, NUM_BAG_SLOTS do
-    for slot = 1, GetContainerNumSlots(bag) do
-      local _, itemCount, locked, _, _, _, _, _, _, itemID = GetContainerItemInfo(bag, slot)
-      if itemID and not locked then
-        --local itemName, _, _, _, _, _, _, itemStackCount = GetItemInfo(itemID)
-        local itemInfo = RS.GetItemInfo(itemID)
-        if itemInfo ~= nil
-            and itemInfo.itemName == item.itemName
-            and itemCount + amountOnMouse <= itemInfo.itemStackCount then
-          PickupContainerItem(bag, slot)
-        end
-      end
-    end
-  end
-
-  -- for all bags: backpack, 1,2,3,4
-  -- Try find a free bag slot
-  for bag = BACKPACK_CONTAINER, NUM_BAG_SLOTS do
-    local numberOfFreeSlots, bagType = GetContainerNumFreeSlots(bag)
+  ---------------------------------------
+  -- Try find a bag which has free space
+  ---------------------------------------
+  for _, bag in ipairs(srcBags) do
+    local numberOfFreeSlots, _bagType = GetContainerNumFreeSlots(bag)
     if numberOfFreeSlots > 0 then
-      local currentBag = bag + 19
-
-      if bag == BACKPACK_CONTAINER then
-        PutItemInBackpack()
-        return
-      else
-        PutItemInBag(currentBag)
-        return
-      end
+      local tmpBag, tmpSlot = rsDropCursorItemIntoBag(dropItem, bag)
+      UseContainerItem(tmpBag, tmpSlot) -- move to the opposite container group bank-bag, or bag-bank
+      return
     end
   end
 end
 
----If need to move items out of bag
----And small stacks below stacksize are found, move entire small stacks
+---Try move full stacks in or out of bank, if not possible, then try move 1 item at a time.
 ---@param state BankRestockCoroState
-local function coroSendPartialStacksToBank(state)
+local function coroBankExchange(state)
+  ----------------------------------------------------
   -- for all bags: backpack, 1,2,3,4 in reverse order
-  -- First see if there are incomplete stacks which entirely can move
-  for bag = NUM_BAG_SLOTS, BACKPACK_CONTAINER, -1 do
+  -- Move excess bag items out to the bank
+  ----------------------------------------------------
+  for _, bag in ipairs(PLAYER_BAGS_REVERSED) do
     for slot = GetContainerNumSlots(bag), 1, -1 do
       local _icon, slotCount, slotLocked, _, _, _, slotItemLink, _, _, slotItemId = GetContainerItemInfo(bag, slot)
       local itemName = slotItemLink and string.match(slotItemLink, "%[(.*)%]")
-      -- local itemInfo = RS.GetItemInfo(itemName)
-      local itemStackCount = 20
 
-      if not slotLocked
-          and state.task[itemName] ~= nil -- this item is in task list
-          and state.task[itemName] < 0 -- negative - move from bag to bank
-          and (itemStackCount and slotCount < itemStackCount) -- and less than full stack
-          and slotCount <= math.abs(state.task[itemName]) -- and not too much
+      -- negative - move from bag to bank
+      if state.task[itemName] and not slotLocked and state.task[itemName] < 0
       then
-        UseContainerItem(bag, slot)
-        state.task[itemName] = state.task[itemName] + slotCount
-        RS.Dbg(string.format("Moved incomplete stack of %s (%d) from bag",
-            itemName, slotCount))
-      end -- if item we should get and its not locked
+        local needToMove = math.abs(state.task[itemName])
+
+        if slotCount <= needToMove then
+          -- If slot does not contain more than the task, take entire slot
+          UseContainerItem(bag, slot)
+          state.task[itemName] = state.task[itemName] + slotCount -- add
+          return
+
+        else
+          local itemInfo = RS.GetItemInfo(slotItemId)
+
+          SplitContainerItem(bag, slot, needToMove)
+          --rsPutSplitItemIntoBags(itemInfo, 1, PLAYER_BAGS, BANK_BAGS)
+          rsSplitSwapCursorItem(itemInfo, PLAYER_BAGS)
+
+          state.task[itemName] = state.task[itemName] + needToMove -- add
+        end
+      end -- if we move to bank
     end -- for slot
-  end -- for bag
-end
+  end -- for player bags
 
----If need to move items out of bag
----And full stacks are found, move entire full stacks
----@param state BankRestockCoroState
-local function coroSendFullStacksToBank(state)
-  -- for all bags: backpack, 1,2,3,4 in reverse order
-  -- First see if there are incomplete stacks which entirely can move
-  for bag = NUM_BAG_SLOTS, BACKPACK_CONTAINER, -1 do
-    for slot = GetContainerNumSlots(bag), 1, -1 do
-      local _icon, slotCount, slotLocked, _, _, _, slotItemLink, _, _, slotItemId = GetContainerItemInfo(bag, slot)
-      local itemName = slotItemLink and string.match(slotItemLink, "%[(.*)%]")
-      -- local itemInfo = RS.GetItemInfo(itemName)
-      local itemStackCount = 20
-
-      if not slotLocked
-          and state.task[itemName] ~= nil -- this item is in task list
-          and state.task[itemName] < 0 -- negative - move from bag to bank
-          and (itemStackCount and slotCount == itemStackCount) -- is full stack
-          and slotCount <= math.abs(state.task[itemName]) -- and not too much
-      then
-        UseContainerItem(bag, slot)
-        state.task[itemName] = state.task[itemName] + slotCount
-        RS.Dbg(string.format("Moved incomplete stack of %s (%d) from bag",
-            itemName, slotCount))
-      end -- if item we should get and its not locked
-    end -- for slot
-  end -- for bag
-end
-
----If need to move items out of bag
----Split stacks to give desired amount and move it.
----@param state BankRestockCoroState
-local function coroSplitStacksToBank(state)
-  -- for all bags: backpack, 1,2,3,4 in reverse order
-  -- First see if there are incomplete stacks which entirely can move
-  for bag = NUM_BAG_SLOTS, BACKPACK_CONTAINER, -1 do
-    for slot = GetContainerNumSlots(bag), 1, -1 do
-      local _icon, slotCount, slotLocked, _, _, _, slotItemLink, _, _, slotItemId = GetContainerItemInfo(bag, slot)
-      local itemName = slotItemLink and string.match(slotItemLink, "%[(.*)%]")
-      -- local itemInfo = RS.GetItemInfo(itemName)
-      local itemStackCount = 20
-
-      if not slotLocked
-          and state.task[itemName] ~= nil -- this item is in task list
-          and state.task[itemName] < 0 -- negative - move from bag to bank
-          and (itemStackCount and slotCount == itemStackCount) -- is full stack
-          and slotCount <= math.abs(state.task[itemName]) -- and not too much
-      then
-        UseContainerItem(bag, slot)
-        state.task[itemName] = state.task[itemName] + slotCount
-        RS.Dbg(string.format("Moved incomplete stack of %s (%d) from bag",
-            itemName, slotCount))
-      end -- if item we should get and its not locked
-    end -- for slot
-  end -- for bag
-end
-
----If have too few restock items, and the bank has them, take from bank
----@param state BankRestockCoroState
-local function coroTakeFromBank(state)
-  if state.transferredToBank then
-    return
-  end
-
-  -- full stacks
+  ----------------------------------------------------
+  -- For all bank bags, move needed items back to bags
+  ----------------------------------------------------
   for _, bag in ipairs(BANK_BAGS_REVERSED) do
     for slot = GetContainerNumSlots(bag), 1, -1 do
-      local _, itemCount, locked, _, _, _, itemLink, _, _, itemId = GetContainerItemInfo(bag, slot)
-      local itemName = itemLink and string.match(itemLink, "%[(.*)%]")
+      local _icon, slotCount, slotLocked, _, _, _, slotItemLink, _, _, slotItemId = GetContainerItemInfo(bag, slot)
+      local itemName = slotItemLink and string.match(slotItemLink, "%[(.*)%]")
 
-      if itemId and not locked then
-        local inRestockList = rsIsItemInRestockList(itemName)
+      -- positive - take from bank
+      if state.task[itemName] and not slotLocked and state.task[itemName] > 0
+      then
+        local needToMove = state.task[itemName]
 
-        if not locked and inRestockList then
+        if slotCount <= needToMove then
+          -- If slot does not contain more than the task, take entire slot
+          UseContainerItem(bag, slot)
+          state.task[itemName] = state.task[itemName] - slotCount -- deduct
+          return
 
-          local item = state.currentProfile[rsGetRestockItemIndex(itemName)]
-          local numInBags = state.itemsInBags[item.itemName] or 0
-          local restockNum = item.amount
-          local difference = restockNum - numInBags
+        else
+          -- Move 1 item at a time
+          local itemInfo = RS.GetItemInfo(slotItemId)
 
-          if difference > 0 and itemCount <= difference then
-            UseContainerItem(bag, slot)
+          -- Take 1 drop 1
+          SplitContainerItem(bag, slot, needToMove)
+          --rsPutSplitItemIntoBags(itemInfo, 1, BANK_BAGS, PLAYER_BAGS)
+          rsSplitSwapCursorItem(itemInfo, BANK_BAGS)
 
-            state.itemsInBags[item.itemName] = (state.itemsInBags[item.itemName]
-                and state.itemsInBags[item.itemName] + itemCount or itemCount)
-            state.rightClickedItem = true
-
-            RS.didBankStuff = true
-            --coroutine.yield()
-          end
+          state.task[itemName] = state.task[itemName] - needToMove -- deduct
         end
-      end -- if item we should get and its not locked
+      end
     end -- for slot
-  end -- for bag
-end
-
----@param state BankRestockCoroState
-local function coroSplitStacks(state)
-  if state.rightClickedItem then
-    return
-  end
-
-  -- split stacks
-  for _, bag in ipairs(BANK_BAGS_REVERSED) do
-    for slot = GetContainerNumSlots(bag), 1, -1 do
-      local _, itemCount, locked, _, _, _, itemLink, _, _, itemId = GetContainerItemInfo(bag, slot)
-      local itemName = itemLink and string.match(itemLink, "%[(.*)%]")
-
-      if itemId and not locked then
-        local inRestockList = rsIsItemInRestockList(itemName)
-        local itemInfo = RS.GetItemInfo(itemName)
-
-        if inRestockList then
-          local item = state.currentProfile[rsGetRestockItemIndex(itemName)]
-          local numInBags = state.itemsInBags[item.itemName] or 0
-          local difference = item.amount - numInBags
-
-          -- This will split one stack or move full stack to bags
-          if difference > 0 and itemCount > difference then
-            local toSplit = mod(difference + numInBags, itemInfo.itemStackCount)
-
-            if toSplit == 0 then
-              -- if the amount we need creates a full stack in the inventory we simply have to
-              -- pick up the item and place it on the incomplete stack in our inventory
-              -- if we split stacks here we get an error saying "couldn't split those items."
-              PickupContainerItem(bag, slot)
-            else
-              -- if the amount of items we need doesn't create a full stack then we split
-              -- the stack in the bank and merge it with the one in our inventory.
-              SplitContainerItem(bag, slot, toSplit)
-            end
-
-            rsPutSplitItemIntoBags(item, toSplit)
-
-            RS.didBankStuff = true
-            state.itemsInBags[item.itemName] = (state.itemsInBags[item.itemName]
-                and state.itemsInBags[item.itemName] + toSplit or toSplit)
-
-            state.hasSplitItems = true
-            --coroutine.yield()
-          end
-        end
-      end -- if item we should get and its not locked
-    end -- for slot
-  end -- for bag
+  end -- for bank bags
 end
 
 ---@class BankRestockCoroState
@@ -364,15 +266,13 @@ local function rsCountMoveItems(state)
     if rs.amount < haveInBackpack then
       -- Negative for take from bag, positive for take from bank
       task[rs.itemName] = rs.amount - haveInBackpack
-      RS.Dbg(string.format("To bank %s=%d", rs.itemName, task[rs.itemName]))
     else
-      -- if need more than have in backpack BUT have in bank
+      -- if have in bank
       if rs.amount > haveInBackpack and haveInBank > 0 then
         -- Negative for take from bag, positive for take from bank
         task[rs.itemName] = math.min(
             rs.amount - haveInBackpack, -- don't move more than we need
             haveInBank) -- but not more than have in bank
-        RS.Dbg(string.format("From bank %s=%d", rs.itemName, task[rs.itemName]))
       end
     end
   end -- for all items in restock list
@@ -392,15 +292,7 @@ local function coroutineBankLoadUnload()
                   task              = {} }
 
   state.task = rsCountMoveItems(state)
-
-  -- Now try move small partial stacks first by using them
-  coroSendPartialStacksToBank(state)
-  -- Then try move full stacks by using them
-  coroSendFullStacksToBank(state)
-  -- Then try split and move stack fragments
-  coroSplitStacksToBank(state)
-  --coroTakeFromBank(state)
-  --coroSplitStacks(state)
+  coroBankExchange(state)
 
   if not state.rightClickedItem
       and not state.transferredToBank
@@ -412,6 +304,7 @@ local function coroutineBankLoadUnload()
   end
 end
 
+RS.xxx = coroutineBankLoadUnload -- for debugging restocking coroutin
 restockerCoroutine = coroutine.create(coroutineBankLoadUnload)